requirements.txt
Flask
reportlab
gunicorn
psycopg2-binary
python-dotenv